Move header files into include/midifile folder.
Effectively we add a folder "namespace" to the midifile headers.
This is standard practice when structuring a library (e.g. note
boost does this). When a parent project includes this library it
will be more clear which headers belong to the `midifile`
library.

Also a minor CMake fix is made which will associate the header
folder with the `midifile` CMake target in accordance with modern
CMake best practices.

These changes make the project more friendly to the common
workflow of incorporating the library into a larger project as a
git submodule.
