+++ include/midifile/MidiEventList.h
@@ -13,7 +13,7 @@
 #ifndef _MIDIEVENTLIST_H_INCLUDED
 #define _MIDIEVENTLIST_H_INCLUDED
 
-#include "MidiEvent.h"
+#include "midifile/MidiEvent.h"
 #include <vector>
 
 namespace smf {
