+++ include/midifile/MidiMessage.h
