+++ tools/createmidifile2.cpp
@@ -9,7 +9,7 @@
 //                with convenience functions.
 //
 
-#include "MidiFile.h"
+#include "midifile/MidiFile.h"
 #include <iostream>
 
 using namespace std;
