+++ tools/durations.cpp
@@ -11,8 +11,8 @@
 //                events) will have a duration of 0.
 //
 
-#include "Options.h"
-#include "MidiFile.h"
+#include "midifile/Options.h"
+#include "midifile/MidiFile.h"
 #include <iostream>
 
 using namespace std;
