+++ tools/maxtick.cpp
@@ -9,8 +9,8 @@
 // Description:   Calculate the maximum timestamp in a MIDI file.
 //
 
-#include "MidiFile.h"
-#include "Options.h"
+#include "midifile/MidiFile.h"
+#include "midifile/Options.h"
 
 #include <iostream>
