+++ tools/mid2hex.cpp
@@ -9,8 +9,8 @@
 //                or hex-byte MIDI file) into a hex byte-code MIDI file.
 //
 
-#include "MidiFile.h"
-#include "Options.h"
+#include "midifile/MidiFile.h"
+#include "midifile/Options.h"
 #include <iostream>
 
 using namespace std;
