+++ tools/mid2mtb.cpp
@@ -10,8 +10,8 @@
 //                text file.
 //
 
-#include "MidiFile.h"
-#include "Options.h"
+#include "midifile/MidiFile.h"
+#include "midifile/Options.h"
 
 #include <ctype.h>
 #include <string.h>
