+++ tools/midi2beep.cpp
@@ -10,8 +10,8 @@
 //                on a single channel playing a single note at a time)
 //
 
-#include "MidiFile.h"
-#include "Options.h"
+#include "midifile/MidiFile.h"
+#include "midifile/Options.h"
 
 #include <iostream>
 #include <cmath>
