+++ tools/midi2binasc.cpp
@@ -12,8 +12,8 @@
 //                   http://binasc.googlecode.com
 //
 
-#include "MidiFile.h"
-#include "Options.h"
+#include "midifile/MidiFile.h"
+#include "midifile/Options.h"
 
 #include <iostream>
 #include <iomanip>
