+++ tools/midi2skini.cpp
@@ -11,8 +11,8 @@
 //                SKINI data format.
 //
 
-#include "MidiFile.h"
-#include "Options.h"
+#include "midifile/MidiFile.h"
+#include "midifile/Options.h"
 #include <iostream>
 #include <iomanip>
