+++ tools/midi2text.cpp
@@ -10,8 +10,8 @@
 // Description:   Description: Converts a MIDI file into a text based notelist.
 //
 
-#include "MidiFile.h"
-#include "Options.h"
+#include "midifile/MidiFile.h"
+#include "midifile/Options.h"
 
 #include <ctype.h>
 #include <string.h>
