+++ tools/midimixup.cpp
@@ -9,8 +9,8 @@
 //                into a random order.
 //
 
-#include "MidiFile.h"
-#include "Options.h"
+#include "midifile/MidiFile.h"
+#include "midifile/Options.h"
 #include <iostream>
 #include <random>
 #include <algorithm>
