+++ tools/perfid.cpp
@@ -10,8 +10,8 @@
 //                it is step edit.
 //
 
-#include "MidiFile.h"
-#include "Options.h"
+#include "midifile/MidiFile.h"
+#include "midifile/Options.h"
 #include <iostream>
 #include <iomanip>
 #include <vector>
