+++ tools/readstatus.cpp
@@ -8,8 +8,8 @@
 // Description:   Demonstration of checking the read status.
 //
 
-#include "MidiFile.h"
-#include "Options.h"
+#include "midifile/MidiFile.h"
+#include "midifile/Options.h"
 #include <iostream>
 
 using namespace std;
