+++ tools/redexpress.cpp
@@ -47,8 +47,8 @@
 //
 ///////////////////////////////////////////////////////////
 
-#include "MidiFile.h"
-#include "Options.h"
+#include "midifile/MidiFile.h"
+#include "midifile/Options.h"
 #include <vector>
 #include <iostream>
