+++ tools/retick.cpp
@@ -11,8 +11,8 @@
 //                time values the same as before.
 //
 
-#include "MidiFile.h"
-#include "Options.h"
+#include "midifile/MidiFile.h"
+#include "midifile/Options.h"
 #include <iostream>
 
 using namespace std;
