+++ tools/smfdur.cpp
@@ -9,8 +9,8 @@
 // Description:   Calcualte the total duration of a MIDI file.
 //
 
-#include "MidiFile.h"
-#include "Options.h"
+#include "midifile/MidiFile.h"
+#include "midifile/Options.h"
 #include <iostream>
 
 using namespace std;
