+++ tools/stretch.cpp
@@ -17,8 +17,8 @@
 //                bars, and also sometimes does not preserve
 //                the tempo of the original.
 //   
-#include "MidiFile.h"
-#include "Options.h"
+#include "midifile/MidiFile.h"
+#include "midifile/Options.h"
 #include <iostream>
 #include <cmath>
 #include <algorithm>
