+++ tools/toascii.cpp
@@ -11,8 +11,8 @@
 //                back into the binary format without information loss).
 //
 
-#include "Options.h"
-#include "MidiFile.h"
+#include "midifile/Options.h"
+#include "midifile/MidiFile.h"
 #include <iostream>
 
 using namespace std;
