+++ tools/tobinary.cpp
@@ -12,8 +12,8 @@
 //                information loss).
 //
 
-#include "Options.h"
-#include "MidiFile.h"
+#include "midifile/Options.h"
+#include "midifile/MidiFile.h"
 #include <iostream>
 
 using namespace std;
