+++ tools/type0.cpp
@@ -10,8 +10,8 @@
 //                a type=0 (single track) Standard MIDI file.
 //
 
-#include "Options.h"
-#include "MidiFile.h"
+#include "midifile/Options.h"
+#include "midifile/MidiFile.h"
 #include <iostream>
 
 using namespace std;
