+++ visual-studio/binasc.vcxproj
@@ -41,13 +41,13 @@
   <PropertyGroup Label="UserMacros" />
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">
     <LinkIncremental>true</LinkIncremental>
-    <IncludePath>$(VC_IncludePath);$(WindowsSDK_IncludePath);../include</IncludePath>
+    <IncludePath>$(VC_IncludePath);$(WindowsSDK_IncludePath);../include/midifile</IncludePath>
     <IntDir>$(Configuration)Exe\</IntDir>
     <OutDir>$(SolutionDir)$(Configuration)Exe\</OutDir>
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">
     <LinkIncremental>false</LinkIncremental>
-    <IncludePath>$(VC_IncludePath);$(WindowsSDK_IncludePath);../include</IncludePath>
+    <IncludePath>$(VC_IncludePath);$(WindowsSDK_IncludePath);../include/midifile</IncludePath>
     <IntDir>$(Configuration)Exe\</IntDir>
     <OutDir>$(SolutionDir)$(Configuration)Exe\</OutDir>
   </PropertyGroup>
